Convert tar_extract to a class
